alpha_mvp-main2/requirements.txt
streamlit>=1.28.0
supabase>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
altair>=5.0.0
python-dotenv>=1.0.0
